feat(export): separate running the app from boostrapping it
this will allow apps to import achievibit and use it as a sub path
